+++ board-server/src/main/java/net/stzups/board/Board.java
@@ -6,63 +6,21 @@ import net.stzups.board.config.ConfigBuilder;
 import net.stzups.board.config.configs.ArgumentConfig;
 import net.stzups.board.config.configs.EnvironmentVariableConfig;
 import net.stzups.board.config.configs.PropertiesConfig;
-import net.stzups.board.data.TokenGenerator;
-import net.stzups.board.data.database.flatfile.FlatFileStorage;
-import net.stzups.board.data.objects.Document;
-import net.stzups.board.data.objects.HttpSession;
-import net.stzups.board.data.objects.User;
-import net.stzups.board.data.objects.UserSession;
+import net.stzups.board.data.database.Database;
+import net.stzups.board.data.database.runtime.RuntimeDatabase;
 import net.stzups.board.server.Server;
 
-import java.io.IOException;
+import java.security.SecureRandom;
+import java.util.Random;
 import java.util.logging.Logger;
 
 public class Board {
     private static Logger logger;
     private static Config config;
+    private static final SecureRandom secureRandom = new SecureRandom();
+    private static final Random random = new Random();
 
-    private static final String DEFAULT_DOCUMENT_NAME = "Untitled Document";
-
-    private static FlatFileStorage<Long, User> users;//user id -> user
-    private static FlatFileStorage<Long, Document> documents;//document id -> document
-    private static FlatFileStorage<Long, HttpSession> httpSessions;//http session id -> http session todo unused
-    private static FlatFileStorage<Long, UserSession> userSessions;//user session id -> user session
-
-    public static Document getDocument(long id) {
-        return documents.get(id);
-    }
-
-
-    public static void addUser(User user) {
-        users.put(user.getId(), user);
-    }
-
-    public static User getUser(long id) {
-        return users.get(id);
-    }
-
-    public static UserSession removeUserSession(long token) {
-        return userSessions.remove(token);
-    }
-
-    public static void addUserSession(UserSession userSession) {
-        userSessions.put(userSession.getToken(), userSession);
-    }
-
-    public static HttpSession getHttpSession(long token) {
-        return httpSessions.get(token);
-    }
-
-    public static void addHttpSession(HttpSession httpSession) {
-        httpSessions.put(httpSession.getToken(), httpSession);
-    }
-
-    public static Document createDocument(User owner) {
-        Document document = new Document(TokenGenerator.getSecureRandom().nextLong(), owner, DEFAULT_DOCUMENT_NAME);
-        documents.put(document.getId(), document);
-        return document;
-    }
-
+    private static Database database;//user id -> user
 
     public static void main(String[] args) throws Exception {
         logger = LogFactory.getLogger("Board Server");
@@ -77,18 +35,10 @@ public class Board {
                 .addConfig(new EnvironmentVariableConfig("board."))
                 .build();
 
+        database = new RuntimeDatabase();
+
         Server server = new Server();
         ChannelFuture channelFuture = server.start();
-
-        try {
-            users = new FlatFileStorage<>("users");
-            documents = new FlatFileStorage<>("documents");
-            httpSessions = new FlatFileStorage<>("httpSessions");
-            userSessions = new FlatFileStorage<>("userSessions");
-        } catch (IOException e) {
-            e.printStackTrace();
-            return;
-        }
 
         logger.info("Started Board server in " + (System.currentTimeMillis() - start) + "ms");
 
@@ -109,5 +59,17 @@ public class Board {
 
     public static Config getConfig() {
         return config;
+    }
+
+    public static Database getDatabase() {
+        return database;
+    }
+
+    public static SecureRandom getSecureRandom() {
+        return secureRandom;
+    }
+
+    public static Random getRandom() {
+        return random;
     }
 }