+++ board-server/src/main/java/net/stzups/board/data/TokenGenerator.java
@@ -5,14 +5,5 @@ import java.util.Random;
 
 public class TokenGenerator {
 
-    private static final SecureRandom secureRandom = new SecureRandom();
-    private static final Random random = new Random();
 
-    public static SecureRandom getSecureRandom() {
-        return secureRandom;
-    }
-
-    public static Random getRandom() {
-        return random;
-    }
 }