+++ board-server/src/main/java/net/stzups/board/server/WebSocketInitializer.java
@@ -19,7 +19,7 @@ public class WebSocketInitializer extends ChannelInboundHandlerAdapter {
         if (event instanceof WebSocketServerProtocolHandler.HandshakeComplete) {
             WebSocketServerProtocolHandler.HandshakeComplete handshakeComplete = (WebSocketServerProtocolHandler.HandshakeComplete) event;
             ctx.channel().attr(HTTP_SESSION_KEY).set(HttpSession.getSession(handshakeComplete.requestHeaders(), ((InetSocketAddress) ctx.channel().remoteAddress()).getAddress()));
-            ctx.pipeline().addLast(new PacketHandler());
+            ctx.pipeline().addLast(new PacketHandler());//todo give this a different executor https://stackoverflow.com/questions/49133447/how-can-you-safely-perform-blocking-operations-in-a-netty-channel-handler
         }
     }
 }